moved convert to GenTreePostprocessPass, removed check for types in array_merge_spread and removed unused test
