+++ compiler/pipes/final-check.cpp
@@ -469,8 +469,6 @@ void FinalCheckPass::check_op_func_call(VertexAdaptor<op_func_call> call) {
                  fmt_format("is_null() will be always false for {}", arg_type->as_human_readable()));
     } else if (vk::string_view{function_name}.starts_with("rpc_tl_query")) {
       G->set_untyped_rpc_tl_used();
-    } else if (function_name == "array_merge_spread") {
-      check_array_merge_spread_call(call);
     }
 
     // TODO: express the array<Comparable> requirement in functions.txt and remove these adhoc checks?
@@ -485,19 +483,6 @@ void FinalCheckPass::check_op_func_call(VertexAdaptor<op_func_call> call) {
   }
 
   check_func_call_params(call);
-}
-
-void FinalCheckPass::check_array_merge_spread_call(VertexAdaptor<op_func_call> call) {
-  const auto args = call->args();
-
-  for (const auto &arg : args) {
-    const auto *arg_type = tinf::get_type(arg);
-    if (!arg_type) {
-      continue;
-    }
-
-    kphp_error(arg_type->ptype() == tp_array, fmt_format("Only arrays can be unpacked with spread (...$a) operator, but passed type '{}'", arg_type->as_human_readable()));
-  }
 }
 
 // Inspection: static-var should be initialized at the declaration (with the exception of tp_mixed).
