+++ tests/phpt/spread/001_spread.php
